+++ includes/base/admin-menu/admin-menu.php
@@ -3,6 +3,9 @@
 /* Change menu order */
 add_action( 'admin_menu', 'ejo_change_menu_order' );
 
+/* Remove Tools submenu */
+add_action( 'admin_menu', 'ejo_remove_tools_submenu', 99 );
+
 /* Remove unnecesary menus */
 function ejo_change_menu_order()
 {
@@ -13,4 +16,11 @@ function ejo_change_menu_order()
 		$menu[6] = $menu[5]; 
 	$menu[5] = $menu[20];
 	unset($menu[20]);
+}
+
+
+/* Remove Tools Submenu */
+function ejo_remove_tools_submenu()
+{
+    remove_submenu_page( 'tools.php', 'tools.php' );
 }