+++ includes/module-manager/modules/blog-comments.php
@@ -53,7 +53,6 @@ if ( EJO_Base_Module::is_inactive('blog-comments') ) {
     add_action('admin_init', function() {
 
         remove_meta_box( 'dashboard_recent_comments', 'dashboard', 'normal' );
-        remove_meta_box( 'dashboard_activity', 'dashboard', 'normal');
     });
 
     //* Remove widget