+++ CHANGES.md
@@ -1,3 +1,5 @@
+- Remove WordPress SEO dashboard widget
+
 1.7
 - Added dashboard widget manager
 - Hide blog and comments and forbid access instead of changing post-capabilities. Changing post capabilities conflicts with editing attachment meta and presumably more.